pyproject.toml
[tool.poetry]
name = "huggingface-workflows"
version = "0.1.0"
description = "Custom Pipelines and Data Workflows - From User to Architect"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.12"
transformers = "^4.40.0"
torch = "^2.2.0"
torchvision = "^0.17.0"
torchaudio = "^2.2.0"
datasets = "^2.14.0"
diffusers = "^0.25.0"
accelerate = "^0.25.0"
bitsandbytes = "^0.41.0"
peft = "^0.8.0"
pillow = "^10.2.0"
soundfile = "^0.12.1"
matplotlib = "^3.8.0"
seaborn = "^0.13.0"
# sentencepiece = "^0.1.99"  # Optional - install separately if needed
protobuf = "^4.25.0"  # Add protobuf for tokenizers
python-dotenv = "^1.0.0"
pandas = "^2.1.0"
numpy = "^1.26.0"
requests = "^2.31.0"
tqdm = "^4.66.0"
onnx = "^1.15.0"
optimum = "^1.16.0"
onnxruntime = "^1.16.0"
psutil = "^5.9.0"
gputil = "^1.4.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
black = "^24.0.0"
ruff = "^0.6.0"
jupyter = "^1.0.0"
ipykernel = "^6.29.0"
ipywidgets = "^8.1.0"
memory-profiler = "^0.61.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py312']

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "C4", "SIM"]
